pyproject.toml
[tool.poetry]
name = "rag-documentation-helper"
version = "0.1.0"
description = ""
authors = ["pedro.grilo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-openai = "^0.3.11"
langchain-community = "^0.3.20"
langchainhub = "^0.1.21"
python-dotenv = "^1.1.0"
langchain-milvus = "^0.1.8"
streamlit = "^1.44.0"
beautifulsoup4 = "^4.13.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
